FPGA Mezzanien Library
This C library provides functions based around libmraa and the official
FPGA Mezzanine firmware in order to toggle and setup GPIO pins on the
Arduino and Raspberry Pi headers.

This is still a WIP:
- Arduino header is fully functioning
- RPi header is still WIP

All Documentation will be in the 96Boards documentation repository
since this is technically not a "project".

The documentation page will be linked to README.md once it has been created.